API: New interface to ChildrenReceiver
With nodes that create new instances of themselves by patching themselves, it is impossible for the user to provide a stream of a list of nodes unless those nodes are completely static with no streams attached. So instead I created DynamicNodeList which is like a managed stream of lists of nodes.