API: Remove auto-distinct DOM updates, change cls from prop to attr, etc.
Automatically applying a .distinct filter to DOM updaters is not safe because it does not work as expected when the DOM is updated from multiple sources (e.g. two streams feeding the same attribute, or some external changes affecting the children. The correct way to do this would be to distinct against the current value in the DOM, as opposed to the last emitted value from the stream, but reading the current value from the DOM is actually more expensive than updating it blindly (verified with benchmarks), so this approach makes no sense.

I changed cls from prop to attr because composite keys require reading current value from the DOM, and for
props, that can fail on form elements: https://github.com/raquo/Laminar/issues/136 (there is no performance
penalty for this according to my benchmarks)