API: Make tags and most keys vals instead of lazy vals
Starting with Scala.js 1.11.0, unused fields can now be dead-code-eliminated:
https://www.scala-js.org/news/2022/09/15/announcing-scalajs-1.11.0/

This is pretty limited, so we can not enable this for SvgAttr and some other types, probably because they are not pure enough (for example, SvgAttr creates a `val name` string in its constructor, and its namespaceUri method triggers evaluation of SvgAttr companion object.

I checked with source-map-explorer that the bundle size only increased by about 500 bytes, and that unused keys are not included in the fullOpt bundle. I think the bundle size increase is probably due to the inclusion of unused codecs that scala.js now has to evaluate – a small price to pay, anyway.