API: Remove willSetParent
Not sure if this is a good idea. This makes unmount callbacks execute after the component has unmounted - so they do not have access to the previous parent, and all the element subscriptions are already dead at this point.

Originally I wanted to try this to simplify code and reduce the amount of computations, but I think the benefit is tiny, and the cost is not worth it.