Naming: ChildrenSeq -> laminar.Seq
Hard to find a good name for this. It is an internal Laminar type, so I hope that it is not a big deal that it uses the same name as a Scala collection type. Internally, I always refer to it as `laminar.Seq` instead of just `Seq` to disambiguate it from the Scala collection type.